Move application code to source directory
In order to make coverage.py work more consistently when run from
various environments (installed package versus local filesystem), I'm
changing it so that all project source lives under source. This
requires that the application be installed via PIP in order for testing
to even work, but forcing this ensures that we're testing against
software that is able to be installed, and coverage.py will always be
able to be certain about what code he is verifying access against.