Use pytest-cache for CI-CD test result cleanup
I'm continually seeing skipped tests in Jenkins CI/CD builds of a
large amount, making it seem as if a large number of tests are being
skipped. This is due to pylama caching. I'm hoping that the
pytest-cache package and its --clearcache option will easily get
me around this, while still allowing for local cache runs on an IDE.