Iterates through MatchingContainers and reports errors
This defines the contract that exists between __main__.main() and
containers.MatchingContainers(). main will initialize
MatchingContainers, catch any NoMatchesError exceptions and print
out a message and when no exceptions occur, will iterate over the
returned instance of MatchingContainers, calling sighup() on each
item that the iterator returns.

The main() side of this contract should now be met, and the skeleton
is present in MatchingContainers() to support the calls made, even
if they currently will raise NotImplementedError on access.