MatchingContainers properly initializes a Docker client
The comment at the top of test_containers.py provides a hacked
method of implementing the required functionality. It was figured
out during an iPython session. I don't plan on using this code
verbatim, and it won't last in the comment forever, but it is a
useful reference point :)

This code includes stub tests for all of the unit tests that I plan
to make pass.