Run hostel-huptainer in a container
huptainerception. This tests out the program while installed in
a container through standard processes (with source contained within
a built Docker image). It spins up a container with an appropriately
named label and then runs hostel-huptainer in another container,
verifying that it is HUPPED by inspecting the initial container's
logs.